requirements.txt
altair==5.3.0
altgraph==0.17.4
attrs==23.2.0
auto-py-to-exe==2.44.1
blinker==1.8.2
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
Eel==0.17.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
protobuf==5.27.2
proxy_tools==0.1.0
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pythonnet==3.0.3
pytz==2024.1
pywebview==5.1
pywin32-ctypes==0.2.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
setuptools==71.1.0
six==1.16.0
smmap==5.0.1
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.4.post2
